h2 as uint8
This doesn't change the logic, but by defining h2 as uint8, and using
uint8 elsewhere, we can work properly with higher bits (and define empty
in it's binary form, instead of having to use the negative value and a
comment). This will also allow us future changes without having to play
with overflows.

Signed-off-by: Oleg Zaytsev <mail@olegzaytsev.com>