Extract simd source to a submodule
Avo documentation requires version to be pinned in order to be used in
production, however there was no reference to avo in the modules here,
so it's not possible to reproduce the build.

This extracts avo source to a submodule, in order to prevent it from
being downloaded as a transitive dependency when dolthub/swiss is
required, and adds a go:generate directive to simplify the build
process.

Signed-off-by: Oleg Zaytsev <mail@olegzaytsev.com>